Refactored FOICreateCase.java into CreateCase.java, so that I could fix the step for creating a random CS case. Added a wait into one of the addCorrespondent methods.
