+++ src/test/java/com/hocs/test/glue/decs/DocumentsStepDefs.java
@@ -2,10 +2,12 @@ package com.hocs.test.glue.decs;
 
 import static jnr.posix.util.MethodName.getMethodName;
 import static net.serenitybdd.core.Serenity.pendingStep;
+import static net.serenitybdd.core.Serenity.setSessionVariable;
 
 import com.hocs.test.pages.decs.BasePage;
 import com.hocs.test.pages.decs.CreateCase;
 import com.hocs.test.pages.decs.CreateCase_SuccessPage;
+import com.hocs.test.pages.decs.Dashboard;
 import com.hocs.test.pages.decs.Documents;
 import io.cucumber.java.en.And;
 import io.cucumber.java.en.Then;
@@ -28,7 +30,7 @@ public class DocumentsStepDefs extends BasePage {
 
     @And("I manage the documents of a new case")
     public void iManageTheDocumentsOfANewCase() {
-        createCase.createCSCaseOfTypeWithoutDocument("CS");
+        createCase.createCSCaseOfRandomType();
         createCaseSuccessPage.goToCaseFromSuccessfulCreationScreen();
         safeClickOn(documents.manageDocumentsLink);
     }
