+++ src/test/java/com/hocs/test/glue/foi/FOICreateCaseStepDefs.java
@@ -1,30 +1,30 @@
 package com.hocs.test.glue.foi;
 
 import com.hocs.test.pages.decs.BasePage;
-import com.hocs.test.pages.foi.FOICreateCase;
+import com.hocs.test.pages.decs.CreateCase;
 import io.cucumber.java.en.And;
 
 public class FOICreateCaseStepDefs extends BasePage {
 
-    FOICreateCase foiCreateCase;
+    CreateCase createCase;
 
     @And("I select how the request was received")
     public void iSelectHowTheRequestWasReceived() {
-        foiCreateCase.selectCorrespondenceInboundChannel();
+        createCase.selectCorrespondenceInboundChannel();
     }
 
     @And("I enter the correspondent details")
     public void iEnterTheCorrespondentDetails() {
-        foiCreateCase.enterCorrespondentDetails();
+        createCase.enterCorrespondentDetails();
     }
 
     @And("I select the {string} FOI topic")
     public void iSelectTheFOITopic(String topic) {
-        foiCreateCase.selectFOITopic(topic);
+        createCase.selectFOITopic(topic);
     }
 
     @And("I enter the Request Question")
     public void iEnterTheRequestQuestion() {
-        foiCreateCase.enterRequestQuestion();
+        createCase.enterRequestQuestion();
     }
 }
